talabahamkorbot/requirements.txt
aiogram==3.15.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
alembic==1.13.0
psycopg2-binary
pandas
openpyxl
google-generativeai
pypdf
python-docx
python-pptx
uvicorn
fastapi
redis
fastapi-cache2
apscheduler
celery
firebase-admin
openai
python-multipart
